[feat] Gremlex.Client: Ping server on a scheduled delay
Neptune and other servers can close web socket connections after a certain idle
timeout. To ensure that connections don't go idle and get disconnected, we've
introduced a scheduled delay. The delay can be configured through the
environment config file. If the delay is 0 (default), then a ping will not be
scheduled.